Fix window startup location when starting in tray & starting minimized
WPF won't respect window startup location if it's set to minimized before being shown.

Can't seem to find a good actual fix so we're working around this by simply not showing the window if it's intended to be in the tray, until it is opened via the tray icon, and if it's set to simply be minimized on startup it will start as a window and then minimize itself.

This is less than ideal, but seems to be a limitation of WPF. I might simply combine the two options in the future to avoid this.