Simplify check for friends window
We no longer need to check for the friends list string as part of the window.

This was initially included because we needed the handle of the parent friends window specifically as in order to reload it we were closing it (by handle) and then calling Steam to reopen it. Now we simply tell Steam to toggle it from offline to online to refresh, and thus this is obsolete.