Use UIAutomation to detect when the friends list opens
When it opens, check for an update to the friends css.

Track whether or not the css is in the process of updating, and if a new file is detected in this time wait until update is complete to process.

This should hopefully eliminate cases where the css would update while Steam is running and the patcher would miss it. Further testing needed.