Use Diagnostics.Process more intelligently
Utilize .FirstOrDefault() from Linq to check process existance.

Get process object with .FirstOrDefault() and utilize .WaitForExit() to determine when Steam is fully shutdown for cache clear method.
(Downside to this is SteamService shuts down later than the rest of the processes, but it doesn't appear to be causing any harm. We can't use .WaitForExit() on SteamService as it runs as Admin)