Major overhaul to improve reliability
Instead of hoping steamchat.com gives us the same version of the css as the client, just grab the latest from both cdns. This still has some unreliability, we transitioned to parsing steamchat precisely because the cdns would sometimes not send the latest version without explicitly including the etag, but now with two cdns this may be more reliable.